[WIP] Added factory functionality, see #25
